Declare added history variables exactly as CAM declares its own
The mdims marker on a Python-owned level field was written as a 64-bit
integer where CAM writes 32-bit, the one way a reader could tell the added
variable from the file's native ones. The history-output gate now records
that the added variable's dimensions, dtype, attribute names, and attribute
types match a CAM field in the same file, and the unit test pins the same.

+++ src/freecam/pi_cam/history_output.py
@@ -472,7 +472,10 @@ class PICAMHistoryStream:
                 dimensions,
             )
             if len(dimensions) == 3:
-                created.mdims = 1
+                # CAM writes this marker as a 32-bit integer; a CDF-5 file
+                # would otherwise carry a 64-bit one for Python-owned fields
+                # alone, which is a difference readers can see.
+                created.mdims = np.int32(1)
             created.units = units or "1"
             created.long_name = long_name or name
             created.cell_methods = self.spec.cell_methods
